Update README.md
The Smart Building Using BLYNK project is an IoT-based home automation system. It uses an ESP32 microcontroller to connect appliances like LEDs and buzzers to the BLYNK app for remote control. Users can turn devices on or off through their smartphone. The system also monitors temperature in real-time and displays data on the serial monitor. This setup enhances convenience and security by enabling remote management. The project is designed and tested on the Wokwi simulation platform. It showcases how IoT and cloud services can simplify smart building automation. This is ideal for learning and prototyping connected home systems.